+++ src/application/archived/math_screen.py
@@ -5,7 +5,7 @@ import tkinter as tk
 from tkinter import ttk
 
 from src.application.objects.question import Question
-from src.application.views import results
+from src.application.archived import results
 
 
 # DIFFERENT TYPES OF PROBLEMS AND IDS:
