+++ src/application/archived/registration.py
