+++ src/application/math_facts.py
@@ -1,9 +1,9 @@
 """
 This is the entry to the Math Facts Application
 """
-from src.application.views import registration
+from src.application.archived import registration
 from src.application.welcome import welcome_description
-from src.application.views.welcome import WelcomeView
+from src.application.archived.welcome import WelcomeView
 import tkinter as tk
 
 if __name__ == '__main__':
